pyproject.toml
[project]
name = "chronocap"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "opencv-python>=4.12.0.88",
    "pydantic>=2.11.9",
    "pyyaml>=6.0.3",
    "schedule>=1.2.2",
]
